Perform clean up of repo
- Remove unused files like hw.js
- Restructure directory for library format
