+++ .travis.yml
@@ -2,7 +2,6 @@ language: node_js
 node_js:
   - node
 
-
 script:
   - npm run test
