+++ packages/core-hello/demo/core_hello.html
@@ -2,7 +2,7 @@
 <html>
 <head>
 	<title>Hello World!</title>
-	<script src="core_hello.js"></script>
+	<script src="../core-hello.js"></script>
 	<link rel="stylesheet" type="text/css" href="core_hello.css">
 	<meta content="text/html;charset=utf-8" http-equiv="Content-Type">
 	<meta content="utf-8" http-equiv="encoding">
