+++ src/index.js
@@ -1,0 +1,3 @@
+// Add all the components in our library here.
+
+//import CoreHello from '../packages/core-hello';
